# Improve performance of prune