… common/flatpak-prune.c