@alexlarsson 2021-04-19
@smcv @mwleeds Can you have a look at this? I did some local performance testing and verification that the locks actually locked. However, before we run this on the flathub repos (which we're going to need soon) it would be good to have some more review of the code.