@alexlarsson 2021-04-19
As a note, the merging of the list-all-objects and prune-unreachable codepaths actually save a lot of memory, which was a problem with the flathub repo, because we ran into out-of-memory issues.