@smcv 2021-04-19
I don't really know the libostree side of things (my Flatpak contributions have mostly been around the sandboxing side) so I'd prefer to defer to someone with more libostree knowledge.