@alexlarsson 2021-04-21
We had issues with out of memory on the flathub server due to holding all the object names of every object in the repo in memory, thus this also has a bunch of focus on how much memory the hashtable of reachable objects take. (I.e. we want a compact representation with little allocation overhead.)