@mwleeds 2021-04-21
I was going to ask why this is being implemented in Flatpak rather than libostree, especially given that it creates `.commitmeta2` files effectively changing the repo format, but the answer is in [this comment](https://github.com/ostreedev/ostree/issues/2338#issuecomment-822352431):

> Some of the ideas from there could probably be picked up in ostree (although the parser generator stuff would add more dependencies). I'm doing this in flatpak because we have a pressing need to get this in so we can actually start pruning the flathub repos before we run out of space...