@alexlarsson 2021-04-23
Ok, I pushed a rebased version with all the issues that are resolved now were fixed, merged into one commit, also adding a test for the prune operation. I think this is basically ready to go, but we want to do a final test on the flathub repo.