@wmanley 2021-04-28
> I created a repo with a lot of small commits to test this. It has 9M objects, and pruning with depth=10 deletes 2M of them.

Can you share the code for creating this repo? I'd quite like to test the performance of https://github.com/wmanley/ostree-oxide/blob/master/ostree-repo/src/prune.rs against it.